+++ examples/ai-functions/src/telemetry/generate-text.ts
@@ -22,10 +22,10 @@ run(async () => {
     experimental_telemetry: {
       isEnabled: true,
       functionId: 'my-awesome-function',
-      metadata: {
-        something: 'custom',
-        someOtherThing: 'other-value',
-      },
+    },
+    context: {
+      something: 'custom',
+      someOtherThing: 'other-value',
     },
   });